HPCC-10826 Make WSWU logic accessible to other projects -class NewWsWorkunit moved to workunitshelpers.hpp -static (in nature) methods grouped and organized in WsWuProcess
