+++ esp/services/ws_workunits/ws_workunitsHelpers.hpp
@@ -24,6 +24,7 @@
 #include "jtime.ipp"
 #include "workunit.hpp"
 #include "hqlerror.hpp"
+#include "dllserver.hpp"
 
 #include <list>
 #include <vector>
@@ -327,5 +328,76 @@ public:
     }
 };
 
+class WsWuProcess
+{
+public:
+    static void setWsWuXmlParameters(IWorkUnit *wu, const char *xml, bool setJobname=false);
+    static void submitWsWorkunit(IEspContext& context, IConstWorkUnit* cw, const char* cluster, const char* snapshot, int maxruntime, bool compile, bool resetWorkflow, bool resetVariables,
+            const char *paramXml=NULL, IArrayOf<IConstNamedValue> *variables=NULL, IArrayOf<IConstNamedValue> *debugs=NULL);
+    static void setWsWuXmlParameters(IWorkUnit *wu, const char *xml, IArrayOf<IConstNamedValue> *variables, bool setJobname=false);
+    static void submitWsWorkunit(IEspContext& context, const char *wuid, const char* cluster, const char* snapshot, int maxruntime, bool compile, bool resetWorkflow, bool resetVariables,
+            const char *paramXml=NULL, IArrayOf<IConstNamedValue> *variables=NULL, IArrayOf<IConstNamedValue> *debugs=NULL);
+    static void copyWsWorkunit(IEspContext &context, IWorkUnit &wu, const char *srcWuid);
+    static void runWsWorkunit(IEspContext &context, StringBuffer &wuid, const char *srcWuid, const char *cluster, const char *paramXml=NULL,
+            IArrayOf<IConstNamedValue> *variables=NULL, IArrayOf<IConstNamedValue> *debugs=NULL);
+    static void runWsWorkunit(IEspContext &context, IConstWorkUnit *cw, const char *srcWuid, const char *cluster, const char *paramXml=NULL,
+            IArrayOf<IConstNamedValue> *variables=NULL, IArrayOf<IConstNamedValue> *debugs=NULL);
+    static IException * noteException(IWorkUnit *wu, IException *e, WUExceptionSeverity level=ExceptionSeverityError);
+    static StringBuffer & resolveQueryWuid(StringBuffer &wuid, const char *queryset, const char *query, bool notSuspended=true, IWorkUnit *wu=NULL);
+    static void runWsWuQuery(IEspContext &context, IConstWorkUnit *cw, const char *queryset, const char *query, const char *cluster, const char *paramXml=NULL);
+    static void runWsWuQuery(IEspContext &context, StringBuffer &wuid, const char *queryset, const char *query, const char *cluster, const char *paramXml=NULL);
+    static void checkAndTrimWorkunit(const char* methodName, StringBuffer& input);
+};
+
+class NewWsWorkunit : public Owned<IWorkUnit>
+{
+public:
+    NewWsWorkunit(IWorkUnitFactory *factory, IEspContext &context)
+    {
+        create(factory, context);
+    }
+
+    NewWsWorkunit(IEspContext &context)
+    {
+        Owned<IWorkUnitFactory> factory = getWorkUnitFactory(context.querySecManager(), context.queryUser());
+        create(factory, context);
+    }
+
+    ~NewWsWorkunit() { if (get()) get()->commit(); }
+
+    void create(IWorkUnitFactory *factory, IEspContext &context)
+    {
+        setown(factory->createWorkUnit(NULL, "ws_workunits", context.queryUserId()));
+        if(!get())
+          throw MakeStringException(ECLWATCH_CANNOT_CREATE_WORKUNIT,"Could not create workunit.");
+        get()->setUser(context.queryUserId());
+    }
+
+    void associateDll(const char *dllpath, const char *dllname)
+    {
+        Owned<IWUQuery> query = get()->updateQuery();
+        StringBuffer dllurl;
+        createUNCFilename(dllpath, dllurl);
+        unsigned crc = crc_file(dllpath);
+        associateLocalFile(query, FileTypeDll, dllpath, "Workunit DLL", crc);
+        queryDllServer().registerDll(dllname, "Workunit DLL", dllurl.str());
+    }
+
+    void setQueryText(const char *text)
+    {
+        if (!text || !*text)
+            return;
+        Owned<IWUQuery> query=get()->updateQuery();
+        query->setQueryText(text);
+    }
+
+    void setQueryMain(const char *s)
+    {
+        if (!s || !*s)
+            return;
+        Owned<IWUQuery> query=get()->updateQuery();
+        query->setQueryMainDefinition(s);
+    }
+};
 }
 #endif
